testsuite/requirements.txt
click==8.1.8
execnet==2.1.1
greenlet==3.1.1
iniconfig==2.1.0
mysql-connector-python==9.2.0
packaging==24.2
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.5
pytest-xdist==3.6.1
python-slugify==8.0.4
selectolax==0.3.28
text-unidecode==1.3
typer-slim==0.15.2
typing_extensions==4.13.0
